// test/integration/test/ports/engines/queries/aggregation/combination_spec/_fixture/decimal/generated/contract.d.ts
export type StorageHash =
  StorageHashBase<'6df294e80c2bf81a495f8523bdb805f90682c00df21e27b5353cdb425219505c'>;
export type ExecutionHash = ExecutionHashBase<string>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Item: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly dec: CodecTypes['pg/numeric@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Item: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly dec: CodecTypes['pg/numeric@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly item: {
      readonly db_dec: CodecTypes['pg/numeric@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly item: {
      readonly db_dec: CodecTypes['pg/numeric@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly item: {
              columns: {
                readonly id: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
                readonly db_dec: {
                  readonly nativeType: 'numeric';
                  readonly codecId: 'pg/numeric@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['id'] };
              uniques: readonly [];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly item: { readonly namespace: 'public' & NamespaceId; readonly model: 'Item' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly Item: {
            readonly fields: {
              readonly id: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
              readonly dec: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/numeric@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'item';
              readonly namespaceId: 'public';
              readonly fields: {
                readonly id: { readonly column: 'id' };
                readonly dec: { readonly column: 'db_dec' };
              };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;