// test/integration/test/ports/engines/queries/aggregation/combination_spec/_fixture/numeric/generated/contract.d.ts
export type StorageHash =
  StorageHashBase<'ee3711eb2445707bca30eeee0aacbca0c8536b8ed10c5cbfd7c21ca7da680dc7'>;
export type ExecutionHash = ExecutionHashBase<string>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Item: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly float: CodecTypes['pg/float8@1']['output'];
      readonly int: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Item: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly float: CodecTypes['pg/float8@1']['input'];
      readonly int: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly item: {
      readonly db_float: CodecTypes['pg/float8@1']['output'];
      readonly db_int: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly item: {
      readonly db_float: CodecTypes['pg/float8@1']['input'];
      readonly db_int: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly item: {
              columns: {
                readonly id: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
                readonly db_float: {
                  readonly nativeType: 'float8';
                  readonly codecId: 'pg/float8@1';
                  readonly nullable: false;
                };
                readonly db_int: {
                  readonly nativeType: 'int4';
                  readonly codecId: 'pg/int4@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['id'] };
              uniques: readonly [];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly item: { readonly namespace: 'public' & NamespaceId; readonly model: 'Item' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly Item: {
            readonly fields: {
              readonly id: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
              readonly float: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/float8@1' };
              };
              readonly int: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/int4@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'item';
              readonly namespaceId: 'public';
              readonly fields: {
                readonly id: { readonly column: 'id' };
                readonly float: { readonly column: 'db_float' };
                readonly int: { readonly column: 'db_int' };
              };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;